requirements.txt
opencv-python
ultralytics
python-dotenv
requests
tqdm
pyyaml
numpy